requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.3.0
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==3.2.0
django-extensions==2.0.6
django-filter==2.0.0
django-nose==1.4.5
django-rest-swagger==2.2.0
djangorestframework==3.7.7
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==2.6.11
MarkupSafe==1.1.1
nose==1.3.7
openapi-codec==1.3.2
psycopg2==2.8.4
psycopg2-binary==2.7.4
PyJWT==1.7.1
pytz==2019.3
PyYAML==3.13
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==0.14.1
whitenoise==4.0
